fix(arborist): use hosted-git-info to correctly parse resolved git urls (#8356)
When custom hosted git url are used to install dependencies it's not
installing that dependency when the url is of
`git+ssh://git@customgit.com:a/b/c.git` format instead of
`git+ssh://git@customgit.com/a/b/c.git`.

#### Cause:
` new URL(resolved)` throws an error which is getting caught and in-turn
results in removal of that node.
This behaviour was working as expected before this
https://github.com/npm/cli/pull/8185 which replaced
`hgi.parseUrl(resoved)` call to `new URL(resolved)`

#### Fix:
keep the `hgi.parseUrl` call to correctly parse the git url as
mentioned/explained in this PR https://github.com/npm/cli/pull/5758

Fixes: https://github.com/npm/cli/issues/8331